Improved the bottom reaching detection.
These values are floating point in some environments.
The scroll position plus the height of the visible area may not exactly match the height of the element.

By changing to the calculation formula introduced on MDN, it will correctly determine that you have reached the bottom of the container element.
https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollHeight#determine_if_an_element_has_been_totally_scrolled