Ensure at least one location per section
Previously, if a book had no text content in a section, then no location
would be generated. In the pathological case of a book with no text
content at all (pictures only), no locations are generated for the
entire book. An example of this is the Open Circuits EPUB from No Starch
Press (2023).

By moving the "empty node" detection after the creation of the initial
opening range, the parsing code ensures that a minimum of one location
is generated per section.

Signed-off-by: Ross Williams <ross@ross-williams.net>